+++ src/oracledb/impl/thick/pool.pyx
@@ -1,5 +1,5 @@
 #------------------------------------------------------------------------------
-# Copyright (c) 2020, 2022, Oracle and/or its affiliates.
+# Copyright (c) 2020, 2023, Oracle and/or its affiliates.
 #
 # This software is dual-licensed to you under the Universal Permissive License
 # (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl and Apache License
@@ -255,7 +255,7 @@ cdef class ThickPoolImpl(BasePoolImpl):
         """
         cdef int value
         if dpiPool_getPingInterval(self._handle, &value) < 0:
-            _raise_from_odpi
+            _raise_from_odpi()
         return value
 
     def get_soda_metadata_cache(self):
@@ -264,7 +264,7 @@ cdef class ThickPoolImpl(BasePoolImpl):
         """
         cdef bint value
         if dpiPool_getSodaMetadataCache(self._handle, &value) < 0:
-            _raise_from_odpi
+            _raise_from_odpi()
         return value
 
     def get_stmt_cache_size(self):
@@ -331,14 +331,14 @@ cdef class ThickPoolImpl(BasePoolImpl):
         Internal method for setting the value of the pool-ping-interval.
         """
         if dpiPool_setPingInterval(self._handle, value) < 0:
-            _raise_from_odpi
+            _raise_from_odpi()
 
     def set_soda_metadata_cache(self, bint value):
         """
         Internal method for enabling or disabling the soda metadata cache.
         """
         if dpiPool_setSodaMetadataCache(self._handle, value) < 0:
-            _raise_from_odpi
+            _raise_from_odpi()
 
     def set_stmt_cache_size(self, uint32_t value):
         """